chore(generator): decompose generator into focused package
Split the 858-line generator.py module into a generator/ package with
focused submodules while preserving the full public symbol surface:

* writer.py         - atomic multi-file write path with rollback
* function_app.py   - marker-based function_app.py registration updates
* json_mutators.py  - host.json / local.settings.json.example mutations
* __init__.py       - public add_*/describe_* orchestration, name
                      normalization, and Jinja rendering

Structural refactor only; no behavior or CLI surface changes.

Part of #163
